Rename file for new rule
