+++ tests/errors/syntax_errors/dict_str_keys.py
